refactor: 💡 add import css at the top, remove all empty css comments
