+++ plugins/css-colocate/index.ts
@@ -1,9 +1,11 @@
 import type { Plugin, ResolvedConfig } from 'vite';
 import path from 'path';
+import fs from 'fs-extra';
 import { preprocessCssModules } from './css-preprocess';
 import { resolveCssModule, loadCssModule } from './virtual-modules';
 import { copyCssFiles } from './css-operations';
-import { injectComponentCss, injectRegularCssImports } from './import-inject';
+import { injectComponentCss } from './import-inject';
+import { getTempDir } from './utils';
 
 interface RegularCssImport {
   sourceFile: string;
@@ -18,10 +20,13 @@ export const cssColocatePlugin = (): Plugin => {
   return {
     name: 'vite-plugin-css-colocate',
     apply: 'build',
-    enforce: 'pre',
+    // No enforce setting - use default hook order
 
     // Preprocess CSS modules before build starts
     async buildStart() {
+      // Clean up any old temp files from previous builds
+      await fs.remove(getTempDir(config.root));
+      // Preprocess CSS modules for this build
       await preprocessCssModules(config.root);
     },
 
@@ -54,6 +59,7 @@ export const cssColocatePlugin = (): Plugin => {
       return null;
     },
 
+    // Inject CSS imports after build completes
     async closeBundle() {
       const outputs = [
         { format: 'esm' as const, ext: 'js' as const },
@@ -68,5 +74,89 @@ export const cssColocatePlugin = (): Plugin => {
     },
   };
 };
+
+async function injectRegularCssImports(
+  regularImports: RegularCssImport[],
+  config: { root: string },
+  format: 'esm' | 'cjs',
+  ext: string
+): Promise<void> {
+  if (regularImports.length === 0) return;
+
+  const distDir = path.join(config.root, 'dist', format);
+  const srcDir = path.join(config.root, 'src');
+
+  for (const { sourceFile, cssPaths } of regularImports) {
+    // Find the corresponding JS output file
+    const relativeToSrc = path.relative(srcDir, sourceFile);
+    const jsOutputFile = path.join(distDir, relativeToSrc.replace(/\.tsx?$/, `.${ext}`));
+
+    if (!(await fs.pathExists(jsOutputFile))) continue;
+
+    let content = await fs.readFile(jsOutputFile, 'utf-8');
+
+    // Build all import statements first
+    const importStatements: string[] = [];
+
+    for (const cssImportPath of cssPaths) {
+      // Resolve the CSS file path
+      let cssSourcePath: string | null = null;
+
+      if (cssImportPath.startsWith('@/')) {
+        cssSourcePath = path.join(srcDir, cssImportPath.slice(2));
+      } else if (cssImportPath.startsWith('./') || cssImportPath.startsWith('../')) {
+        cssSourcePath = path.resolve(path.dirname(sourceFile), cssImportPath);
+      }
+
+      if (!cssSourcePath || !(await fs.pathExists(cssSourcePath))) continue;
+
+      // Calculate output CSS path in dist
+      const cssRelativeToSrc = path.relative(srcDir, cssSourcePath);
+      const cssOutputPath = path.join(distDir, cssRelativeToSrc);
+
+      // Copy CSS file if not already there
+      if (!(await fs.pathExists(cssOutputPath))) {
+        await fs.ensureDir(path.dirname(cssOutputPath));
+        await fs.copy(cssSourcePath, cssOutputPath);
+      }
+
+      // Calculate relative path from JS file to CSS file
+      const cssRelativeToJs = path.relative(path.dirname(jsOutputFile), cssOutputPath);
+      const normalizedCssPath = cssRelativeToJs.replace(/\\/g, '/');
+      const importPath = normalizedCssPath.startsWith('.')
+        ? normalizedCssPath
+        : `./${normalizedCssPath}`;
+
+      importStatements.push(
+        format === 'esm' ? `import "${importPath}";` : `require("${importPath}");`
+      );
+    }
+
+    // Remove all "/* empty css */" comments
+    content = content.replace(/\/\*\s*empty css\s*\*\//g, '');
+
+    // Inject all import statements at the top (after 'use client' directive if present)
+    if (importStatements.length > 0) {
+      const importCode = importStatements.join('\n') + '\n';
+
+      // Find position after 'use client' directive if present
+      let insertPos = 0;
+      const useClientMatch = content.match(/^(['"]use client['"];?)/);
+      if (useClientMatch) {
+        insertPos = useClientMatch[0].length;
+        // Add newline if not present
+        if (content[insertPos] !== '\n') {
+          insertPos = content.indexOf('\n', insertPos) + 1;
+        } else {
+          insertPos++;
+        }
+      }
+
+      content = content.slice(0, insertPos) + importCode + content.slice(insertPos);
+    }
+
+    await fs.writeFile(jsOutputFile, content);
+  }
+}
 
 export default cssColocatePlugin;
