+++ plugins/css-colocate/utils.ts
@@ -11,7 +11,7 @@ export const fileExists = (p: string) =>
 export const findFiles = async (dir: string, pattern: string): Promise<string[]> =>
   glob(pattern, { cwd: dir, absolute: true });
 
-export const getTempDir = (rootDir: string) => path.join(rootDir, '.temp', 'css-modules');
+export const getTempDir = (rootDir: string) => path.join(rootDir, '.css-modules-temp');
 
 export const createImportStatement = (importPath: string, format: 'esm' | 'cjs') =>
   format === 'esm' ? `import "${importPath}";` : `require("${importPath}");`;
